requirements.txt
distro==1.8.0
et-xmlfile==1.1.0
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.1
PyPDF2==3.0.1
PyPDF4==1.27.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tabula-py==2.7.0
tzdata==2023.3
XlsxWriter==3.1.0
